Remove trends-related postgresql code
